Fix Collections list sort order
javascript .sort method returns a sorted list based on string value - this place all items that start with an upper case before all items that start with a lower case. iNaturalist is an example of a collection that would appear to be missing as it would be at or near the bottom of the sorted collections.  The default .sort method also causes issues for some languages.  .localeCompare uses browser settings to determine appropriate sort order.